Tests for Hive V1
